loan_app_project/requirements.txt

Django==3.2.5

djangorestframework==3.14.0

psycopg2-binary==2.9.1

pandas==1.3.3

celery==5.2.2

importlib-metadata==4.13.0

redis==5.0.0

openpyxl==3.0.9

requests==2.26.0